Re-organize Pydantic models and types used in handlers
